Combining the shore and ocean surface builders for better sea ice
